-Store offsets as 64-bit instead of 32-bit -Fix several issues parsing PDB files over 4gb

+++ src/PESpy.PowerShell/GetMissing95Stub.cs
@@ -116,7 +116,7 @@ namespace PESpy.PowerShell
             {
                 if (xref.Kind == XRefKind.To && accessor.TryGetFunctionForAddress(xref.Other, out var exportEntity))
                 {
-                    stubAddrs.Add(exportEntity.TargetAddress);
+                    stubAddrs.Add((int) exportEntity.TargetAddress);
                     stubEntities.Add(exportEntity);
                 }
             }
